+++ voiceflousion/src/integrations/discord/discord_update.rs
@@ -9,7 +9,7 @@ pub struct DiscordUpdate{
     update_base: UpdateBase
 }
 
-impl Deref<Target=UpdateBase> for DiscordUpdate {
+impl Deref for DiscordUpdate {
     type Target = UpdateBase;
 
     fn deref(&self) -> &Self::Target {
@@ -19,6 +19,6 @@ impl Deref<Target=UpdateBase> for DiscordUpdate {
 
 impl Update for DiscordUpdate{
     fn from_request_body(body: Value) -> VoiceflousionResult<Self> {
-        todo!()
+        unimplemented!()
     }
 }