+++ voiceflousion/src/server/mod.rs
@@ -1,7 +1,13 @@
 mod voiceflousion_server;
 pub mod handlers;
 mod endpoints;
+pub mod traits;
+
+#[cfg(not(feature = "advanced"))]
 mod subtypes;
-mod traits;
+
+#[cfg(feature = "advanced")]
+pub mod subtypes;
+
 
 pub use voiceflousion_server::VoiceflousionServer;