+++ voiceflousion/src/server/traits/mod.rs
@@ -1,5 +1,14 @@
 mod bot_handler;
 mod server_client;
+#[cfg(all(not(feature = "advanced"), any(
+    feature = "telegram",
+    feature = "whatsapp",
+    feature = "discord_unimplemented"
+)))]
+mod utils;
+
+#[cfg(feature = "advanced")]
+pub mod utils;
 
 pub use self::bot_handler::BotHandler;
 pub use self::server_client::ServerClient;