+++ voiceflousion/src/server/voiceflousion_server.rs
@@ -2,20 +2,23 @@ use std::collections::HashMap;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use axum::{Extension, Json, Router};
+use axum::extract::{Path, Query};
 use axum::routing::post;
 use serde_json::Value;
 use crate::core::base_structs::ClientsManager;
 use crate::server::endpoints::{get_auth_endpoint, main_endpoint};
+use crate::server::subtypes::{QueryParams, VoiceflousionHeadersWrapper};
 use crate::server::traits::{BotHandler, ServerClient};
 
 /// VoiceflousionServer is responsible for handling HTTP requests to bots and routing them to the appropriate handlers.
 ///
-/// This struct contains bots clients manager, base part of the server's HTTP endpoint URL, and the handler function for processing incoming webhook requests.
+/// This struct contains bots clients manager, optional part of the server's HTTP endpoint URL, and the handler function for processing incoming webhook requests.
+/// It also handles CORS settings to allow or restrict access to specific origins.
 pub struct VoiceflousionServer<C: ServerClient + 'static> {
     /// Manager for handling multiple bots clients.
     clients: Option<Arc<ClientsManager<C >>>,
-    /// Base part of the server's HTTP endpoint URL.
-    base_url: String,
+    /// Optional; extending part of the server's HTTP endpoint URL.
+    extend_url: Option<String>,
     /// Handler function for processing incoming webhook requests.
     handler: Arc<dyn BotHandler<C>>,
     /// Allowed origins for CORS settings stored in a HashMap for fast lookup.
@@ -27,7 +30,6 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
     ///
     /// # Parameters
     ///
-    /// * `base_url` - Base part of the server's HTTP endpoint URL.
     /// * `webhook_handler` - The handler  function for processing incoming webhook requests.
     ///
     /// # Returns
@@ -41,17 +43,17 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
     /// use voiceflousion::server::handlers::base_dialog_handler;
     /// use voiceflousion::server::VoiceflousionServer;
     ///
-    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new("telegram".to_string(), {
+    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new({
     ///             |update, client| Box::pin(base_dialog_handler(update, client))
     ///  });
     /// ```
-    pub fn new(base_url: String, webhook_handler: impl BotHandler<C> + 'static) -> Self {
+    pub fn new(webhook_handler: impl BotHandler<C> + 'static) -> Self {
         let handler =  Arc::new(move |update: C::ClientUpdate<'static>, client: Arc<C>| {
             webhook_handler(update, client)
         });
         Self {
             clients: None,
-            base_url,
+            extend_url: None,
             handler,
             allowed_origins: Arc::new(None)
         }
@@ -84,7 +86,7 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
     ///
     /// let telegram_client_manager = Arc::new(ClientsManager::from_clients(vec![telegram_client]));
     ///
-    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new("telegram".to_string(), {
+    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new({
     ///             |update, client| Box::pin(base_dialog_handler(update, client))
     ///  })
     ///  .set_clients_manager(telegram_client_manager);
@@ -110,7 +112,7 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
     /// use voiceflousion::integrations::telegram::TelegramClient;
     /// use voiceflousion::server::handlers::base_dialog_handler;
     ///
-    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new("telegram".to_string(), {
+    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new({
     ///             |update, client| Box::pin(base_dialog_handler(update, client))
     /// })
     /// .enable_default_origins();
@@ -154,7 +156,7 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
     ///     "http://another-example.com",
     /// ];
     ///
-    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new("telegram".to_string(), {
+    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new({
     ///             |update, client| Box::pin(base_dialog_handler(update, client))
     /// })
     /// .override_allow_origins(additional_origins);
@@ -166,6 +168,63 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
         }
         self.allowed_origins = Arc::new(Some(origins_map));
         self
+    }
+
+    /// Sets an additional extension for the server's URL.
+    ///
+    /// # Parameters
+    ///
+    /// * `extend_url` - The additional string to extend the base URL.
+    ///
+    /// # Returns
+    ///
+    /// The updated `VoiceflousionServer` instance.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use voiceflousion::server::VoiceflousionServer;
+    /// use voiceflousion::integrations::telegram::TelegramClient;
+    /// use voiceflousion::server::handlers::base_dialog_handler;
+    ///
+    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new({
+    ///             |update, client| Box::pin(base_dialog_handler(update, client))
+    /// })
+    /// .set_extend_url("extra");
+    /// ```
+    pub fn set_extend_url(mut self, extend_url: &str) -> Self {
+        self.extend_url = Some(extend_url.to_string());
+        self
+    }
+
+    /// Constructs the route path based on the base URL and optional extension.
+    ///
+    /// # Returns
+    ///
+    /// A `String` representing the route path.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use voiceflousion::server::VoiceflousionServer;
+    /// use voiceflousion::integrations::telegram::TelegramClient;
+    /// use voiceflousion::server::handlers::base_dialog_handler;
+    ///
+    /// let voiceflousion_telegram_server = VoiceflousionServer::<TelegramClient>::new({
+    ///             |update, client| Box::pin(base_dialog_handler(update, client))
+    /// })
+    /// .set_extend_url("extra");
+    ///
+    /// let route = voiceflousion_telegram_server.get_route();
+    /// assert_eq!(route, "/telegram/:id/extra");
+    /// ```
+    pub fn get_route(&self) -> String{
+        if let Some(url) = &self.extend_url{
+            format!("/{}/:id/{}", C::BASE_URL, url)
+        }
+        else{
+            format!("/{}/:id", C::BASE_URL)
+        }
     }
 
     /// Starts the server and begins listening for incoming requests.
@@ -196,7 +255,7 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
     ///     let telegram_client_manager = Arc::new(ClientsManager::from_clients(vec![telegram_client]));
     ///
     ///     tokio::spawn(async move {
-    ///         VoiceflousionServer::<TelegramClient>::new("telegram".to_string(), {
+    ///         VoiceflousionServer::<TelegramClient>::new({
     ///             |update, client| Box::pin(base_dialog_handler(update, client))
     ///         })
     ///         .set_clients_manager(telegram_client_manager)
@@ -206,63 +265,61 @@ impl<C: ServerClient + 'static> VoiceflousionServer<C> {
     /// }
     /// ```
     pub async fn run(self, address: impl Into<SocketAddr>) {
-        let base_url = self.base_url.clone();
-        let router = self.create_router(&base_url).await.into_make_service();
+        let route = self.get_route();
+        let router = self.create_router(route.clone()).await.into_make_service();
 
         // Start the HTTP server
         let ip = address.into();
         let listener = tokio::net::TcpListener::bind(ip).await.unwrap();
 
-        println!("Server is set on {}/{}", ip, base_url);
-        println!("Bots without authentication token are available on {}/{}/<bot_id>", ip, base_url);
-        println!("Bots with authentication token are available on {}/{}/<bot_id>/?token=<token>", ip, base_url);
+        println!("Server is set on {}", ip);
+        println!("Bots without authentication token are available on {}{}", ip, route);
+        println!("Bots with authentication token are available on {}{}/?token=<token>", ip, route);
 
         axum::serve(listener, router).await.unwrap();
     }
 
 
-    /// Creates a new router with the specified base URL.
+    /// Creates a new router for handling HTTP requests.
     ///
     /// # Parameters
     ///
-    /// * `base_url` - The base part of the server's HTTP endpoint URL.
+    /// * `url` - The constructed URL route for the server.
     ///
     /// # Returns
     ///
     /// A `Router` instance configured with the appropriate routes and CORS settings.
-    async fn create_router(self, base_url: &String) -> Router{
+    async fn create_router(self, url: String) -> Router{
         let clients = self.clients.clone().expect("Webhook is not set");
         let handler = self.handler.clone();
         let optional_allowed_origins = self.allowed_origins.clone();
-
         Router::new()
-            .route(&format!("/{}/:id", base_url),
-                   post({
+            .route(&url, post({
                        let clients = clients.clone();
                        let optional_allowed_origins = optional_allowed_origins.clone();
                        let handler = handler.clone();
-                       move |origin_header, path, params, json: Json<Value>| {
+                       move |headers: VoiceflousionHeadersWrapper,  path: Path<String>, params: Query<QueryParams>, json: Json<Value>| {
                            main_endpoint(
                                path,
                                params,
                                json,
+                               headers,
                                Extension(clients),
                                Extension(optional_allowed_origins),
-                               Extension(handler),
-                               origin_header
+                               Extension(handler)
                            )
                        }
                    })
                    .get({
                        let clients = clients.clone();
                        let optional_allowed_origins = optional_allowed_origins.clone();
-                       move |origin_header, path, params| {
+                       move |headers: VoiceflousionHeadersWrapper, path: Path<String>, params: Query<QueryParams>| {
                             get_auth_endpoint(
                                 path,
                                 params,
+                                headers,
                                 Extension(clients),
-                                Extension(optional_allowed_origins),
-                                origin_header
+                                Extension(optional_allowed_origins)
                             )
                        }
                    }),